pyproject.toml
[project]
name = "go-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.1",
    "fastapi>=0.120.3",
    "google-genai>=1.47.0",
    "minio>=7.2.18",
    "moviepy>=1.0.3",
    "numpy>=2.3.4",
    "pillow>=12.0.0",
    "pydantic>=2.12.3",
    "python-decouple>=3.8",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
]
